added flags and removed project connect
